feat(#1470): Drilldown-Tagesfenster folgt der Ortszeit der Tour
PO-Entscheidung 2026-08-03 ("Ja, Ortszeit"), vorgelegt beim Abschluss von
#1465 als Produktfrage.

Bisher richteten sich "heute" und "morgen" nach der Weltzeit. Auf Korsika
begann "morgen" um 02:00 Ortszeit -- die ersten zwei Stunden des Tages
fehlten, die letzten zwei des Vortages waren dabei. Bei einer Tour in
Neuseeland waeren es zwoelf Stunden gewesen. Auch "heute" war betroffen:
day_date war das UTC-Datum, kurz nach Ortsmitternacht meldete der
Drilldown den Vortag. Das stand nicht im Ticket, sondern fiel beim Messen
auf.

Ein gemeinsamer Helfer _day_window() ersetzt zwei wortgleiche Bloecke in
beiden Handlern -- ein Kalender, eine Zonenaufloesung, durchgereicht statt
zweimal geholt.

DREI Funde des Adversary, die den ersten Entwurf gerettet haben:

1. Das Fenster war auf feste 24 Stunden genagelt. An einem Ortstag mit 25
   Stunden fehlte Stunde 23, an einem mit 23 blutete eine Zeile des
   Folgetages herein. Dieser Fehler waere ERST DURCH diese Aenderung
   entstanden -- die Weltzeit kennt keine Umstellung. Jetzt reicht das
   Fenster von Ortsmitternacht bis zur naechsten, die Stundenzahl faellt
   als Differenz an (23/24/25, float wegen Zonen mit halbstuendigem
   Wechsel).

2. Zwei Tests hingen an den Spaltenbeschriftungen statt an den Werten.
   Unter der Rueckmutation blieben sie gruen, obwohl andere Daten
   angezeigt wurden. Die Testdaten tragen jetzt einen ueber die Reihe
   eindeutigen Fingerabdruck.

3. Der Randfall war zehnfach unterschaetzt. Anker war die ERSTE Etappe
   der Tour; bei Neuseeland -> Korsika zehn Stunden daneben. Jetzt die
   Etappe des Weltzeit-Tages -- der Fehler schrumpft auf "Zonenwechsel an
   genau diesem Tag".

Der wertvollste Fund kam beim Bauen: Python ignoriert die Zeitzone, wenn
zwei Zeitpunkte dasselbe tzinfo-Objekt tragen, und rechnet auf den nackten
Wanduhr-Werten -- gemessen 24.0 an JEDEM Tag, auch an den Wechseltagen.
Die naheliegende Fassung haette gruen ausgesehen und nichts behoben. Die
Umrechnung ueber UTC ist deshalb kein Zierrat; Messwerte und eine eigene
Mutation stehen im Docstring von _hours_between.

"heute" bleibt bewusst bei 12 Stunden: das ist eine Dauer ab jetzt, keine
Kalendertagsgrenze, und damit sommerzeit-immun. Ein Kappen an der
Ortsmitternacht wuerde die Vorschau kurz vor Mitternacht auf Minuten
schrumpfen lassen. Begruendung im Docstring.

Adversary: VERIFIED nach 3 Runden. Vier Mutationen mit belegter
Trennschaerfe, jede mit Pruefsummen vor/nach. 57 Tests gruen, breite
Regression 231 gruen.

Nicht enthalten, gemeldet: vier weitere Stellen mit derselben
Verwechslung -- eine davon loest einen VERSAND aus statt nur eine
Anzeige und braucht eine eigene Abwaegung.

+++ src/services/weather_extractor.py
@@ -111,8 +111,12 @@ class WeatherExtractor:
         trip_id: str,
         metric: str,
         from_time: Optional[datetime] = None,
-        hours: int = 12,
+        hours: float = 12,
     ) -> DrilldownResult:
+        # ``hours`` ist float, seit das Tagesfenster ein ORTSTAG ist (#1470):
+        # der hat an Sommerzeit-Wechseln 23 oder 25 Stunden, in Zonen mit
+        # halbstuendigem Wechsel auch 23.5. ``timedelta(hours=...)`` kann das,
+        # ein ``int`` haette es abgeschnitten.
         # Normalisierung an der GRENZE, nicht am Vergleich: `from_time` kommt
         # von aussen (Aufrufer bauen es aus `received_at`) und wird hier einmal
         # auf die Hausnorm gezogen. Ein `if ts.tzinfo` unten in der Fenster-
